+++ http4k-testing/webdriver/src/main/kotlin/org/http4k/webdriver/JSoupElementFinder.kt
@@ -9,7 +9,7 @@ import org.openqa.selenium.internal.FindsByCssSelector
 import org.openqa.selenium.internal.FindsById
 import org.openqa.selenium.internal.FindsByTagName
 
-internal class JSoupElementFinder(private val navigate: Navigate, private val getURL: GetURL, private val element: Element) :
+class JSoupElementFinder(private val navigate: Navigate, private val getURL: GetURL, private val element: Element) :
     FindsByCssSelector, FindsByTagName, FindsById, FindsByClassName, SearchContext {
     override fun findElementByClassName(className: String) = findElementsByClassName(className).firstOrNull()