+++ http4k-testing/webdriver/src/test/kotlin/org/http4k/webdriver/JSoupWebElementTest.kt
@@ -37,7 +37,11 @@ class JSoupWebElementTest {
     private fun form(method: Method = POST) = JSoupWebElement(navigate, getURL, Jsoup.parse("""
         <form method="${method.name}" action="/posted">
             <input id="text" type="text"/>
-            <textarea id="textarea"/>
+            <input id="checkbox" name="checkedCheckbox" type="checkbox" value="checkedCheckbox" checked/>
+            <input id="checkbox" name="uncheckedCheckbox" type="checkbox" value="uncheckedCheckbox"/>
+            <input id="radio" name="radio" type="radio" value="checkedRadio" checked/>
+            <input id="radio" name="radio" type="radio" value="uncheckedRadio"/>
+            <textarea id="textarea" value="bob"/>
             <p>inner</p>
         </form>
         """)).findElement(By.tagName("form"))!!
@@ -116,7 +120,7 @@ class JSoupWebElementTest {
     @Test
     fun `submit an element inside the form`() {
         form(DELETE).findElement(By.tagName("p"))!!.submit()
-        assertThat(newLocation, equalTo(DELETE to "/posted"))
+        assertThat(newLocation, equalTo(DELETE to "/posted?checkedCheckbox=checkedCheckbox&radio=checkedRadio"))
     }
 
     @Test