+++ CHANGELOG.md
@@ -7,6 +7,7 @@ changes with their rationale when appropriate:
 - [http4k-core] Add `Request.source` to provide extra information about the request origin (address/port/scheme). H/T @kam1sh and @jshiell for the contributions.
 - [http4k-security-oauth] Add OAuth provider configuration for Facebook. H/T @knyttl for the PR.
 - [http4k-server-netty] Implement KeepAlive. H/T @carbotaniuman for the PR.
+- [http4k-bom] New Bill-Of-Materials module! 
 
 ### v3.249.0
 - [all] Upgrade some dependency versions.