+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate:
 
-### v3.282.0 (uncut)
+### v3.282.0
 - **http4k-serverless-*** : Tidy implementations to be consistent.
+- **http4k-testing-webdriver-*** : Fixed radio buttons submitting even when not selected.
 
 ### v3.281.0
 - **http4k-*** : Upgrade some dependency versions.